+++ lightningd/opening_control.c
@@ -1273,6 +1273,11 @@ static struct command_result *json_fund_channel_start(struct command *cmd,
 				    "Feerate below feerate floor");
 	}
 
+	if (!topology_synced(cmd->ld->topology)) {
+		return command_fail(cmd, FUNDING_STILL_SYNCING_BITCOIN,
+				    "Still syncing with bitcoin network");
+	}
+
 	peer = peer_by_id(cmd->ld, id);
 	if (!peer) {
 		return command_fail(cmd, LIGHTNINGD, "Unknown peer");
@@ -1363,6 +1368,11 @@ static struct command_result *json_fund_channel(struct command *cmd,
 	if (*feerate_per_kw < feerate_floor()) {
 		return command_fail(cmd, LIGHTNINGD,
 				    "Feerate below feerate floor");
+	}
+
+	if (!topology_synced(cmd->ld->topology)) {
+		return command_fail(cmd, FUNDING_STILL_SYNCING_BITCOIN,
+				    "Still syncing with bitcoin network");
 	}
 
 	peer = peer_by_id(cmd->ld, id);